+++ docs/history.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add `--json` flag to the `env show` command
 - Add environment collector API method `finalize_config`
+- Add more conditions for environment option overrides
 - Expose helper for lazily loading modules
 - Bump the minimum supported version of Hatchling
 - Upgrade `tomlkit` for better formatting